moved db_adapter package; dropped unuused v2 models;
