+++ tests/test_query_hazard_caching.py
@@ -12,7 +12,7 @@ from pynamodb.models import Model
 
 from toshi_hazard_store import model, query
 from toshi_hazard_store.model.caching import cache_store
-from toshi_hazard_store.v2.db_adapter import ensure_class_bases_begin_with
+from toshi_hazard_store.db_adapter import ensure_class_bases_begin_with
 
 HAZARD_MODEL_ID = 'MODEL_THE_FIRST'
 vs30s = [250, 350, 450]
