+++ toshi_hazard_store/db_adapter/dynamic_base_class.py
