+++ toshi_hazard_store/db_adapter/pynamodb_adapter_interface.py
