+++ toshi_hazard_store/db_adapter/sqlite/__init__.py
