+++ toshi_hazard_store/db_adapter/sqlite/sqlite_adapter.py
