+++ toshi_hazard_store/db_adapter/sqlite/sqlite_store.py
