+++ toshi_hazard_store/db_adapter/test/test_adapter_setup.py
