+++ toshi_hazard_store/model/caching/cache_store.py
@@ -7,7 +7,7 @@ from typing import Iterable
 from pynamodb.expressions.condition import Condition
 
 from toshi_hazard_store.config import DEPLOYMENT_STAGE, LOCAL_CACHE_FOLDER
-from toshi_hazard_store.v2.db_adapter.sqlite.sqlite_store import (  # noqa
+from toshi_hazard_store.db_adapter.sqlite.sqlite_store import (  # noqa
     ensure_table_exists,
     execute_sql,
     get_model,
